[fix] parseInt to parseFloat

+++ src/app/graphic/graphic.component.ts
@@ -116,7 +116,7 @@ export class GraphicComponent implements OnInit {
       let lastRegisterTime = allRegisters[allRegisters.length-1]['time'];
       let lastRegisterHigh = allRegisters[allRegisters.length-1]['high'];
       let lastRegisterLow = allRegisters[allRegisters.length-1]['low'];
-      let lastRegisterAverage = parseInt(((lastRegisterHigh + lastRegisterLow) / 2).toFixed(2));
+      let lastRegisterAverage = parseFloat(((lastRegisterHigh + lastRegisterLow) / 2).toFixed(2));
 
       let jsDate = new Date(lastRegisterTime * 1000);
